Introduced expectBadge helper
