+++ services/wercker/wercker.tester.js
@@ -1,24 +1,23 @@
 'use strict'
 
-const Joi = require('joi')
 const { isBuildStatus } = require('../../lib/build-status')
 const t = (module.exports = require('../tester').createServiceTester())
 
 t.create('Build status')
   .get('/build/wercker/go-wercker-api.json')
-  .expectJSONTypes(Joi.object().keys({ name: 'build', value: isBuildStatus }))
+  .expectBadge({ label: 'build', message: isBuildStatus })
 
 t.create('Build status (with branch)')
   .get('/build/wercker/go-wercker-api/master.json')
-  .expectJSONTypes(Joi.object().keys({ name: 'build', value: isBuildStatus }))
+  .expectBadge({ label: 'build', message: isBuildStatus })
 
 t.create('Build status (application not found)')
   .get('/build/some-project/that-doesnt-exist.json')
-  .expectJSON({ name: 'build', value: 'application not found' })
+  .expectBadge({ label: 'build', message: 'application not found' })
 
 t.create('Build status (private application)')
   .get('/build/wercker/blueprint.json')
-  .expectJSON({ name: 'build', value: 'private application not supported' })
+  .expectBadge({ label: 'build', message: 'private application not supported' })
 
 t.create('Build passed (mocked)')
   .get('/build/wercker/go-wercker-api.json?style=_shields_test')
@@ -27,9 +26,9 @@ t.create('Build passed (mocked)')
       .get('/wercker/go-wercker-api/builds?limit=1')
       .reply(200, [{ status: 'finished', result: 'passed' }])
   )
-  .expectJSON({
-    name: 'build',
-    value: 'passing',
+  .expectBadge({
+    label: 'build',
+    message: 'passing',
     color: 'brightgreen',
   })
 
@@ -40,15 +39,15 @@ t.create('Build failed (mocked)')
       .get('/wercker/go-wercker-api/builds?limit=1')
       .reply(200, [{ status: 'finished', result: 'failed' }])
   )
-  .expectJSON({ name: 'build', value: 'failing', color: 'red' })
+  .expectBadge({ label: 'build', message: 'failing', color: 'red' })
 
 t.create('CI status by ID')
   .get('/ci/559e33c8e982fc615500b357.json')
-  .expectJSONTypes(Joi.object().keys({ name: 'build', value: isBuildStatus }))
+  .expectBadge({ label: 'build', message: isBuildStatus })
 
 t.create('CI status by ID (with branch)')
   .get('/ci/559e33c8e982fc615500b357/master.json')
-  .expectJSONTypes(Joi.object().keys({ name: 'build', value: isBuildStatus }))
+  .expectBadge({ label: 'build', message: isBuildStatus })
 
 t.create('CI status by ID (no runs yet)')
   .get('/ci/559e33c8e982fc615500b357.json')
@@ -57,4 +56,4 @@ t.create('CI status by ID (no runs yet)')
       .get('/runs?applicationId=559e33c8e982fc615500b357&limit=1')
       .reply(200, [])
   )
-  .expectJSON({ name: 'build', value: 'not built' })
+  .expectBadge({ label: 'build', message: 'not built' })
